More env config
We need to duplicate a bunch of the variables in different formats
because... well, frankly we don't know why. But if you don't and opt to
only use one set funny stuff happens like the app will happily create a
new instance but then be unable to operate. Like creation and operation
are two indepedent sets of credentials.